edit field

+++ Back-end/users/models.py
@@ -45,7 +45,7 @@ class User(AbstractBaseUser, PermissionsMixin, TrackingModel):
     is_superuser = models.BooleanField(default=False)
     is_verified = models.BooleanField(default=False)
     result=models.CharField(max_length=1000, default=" لم تظهر بعد" )
-    grade = models.IntegerField(default= None)
+    grade = models.FloatField(default= None)
     
     USERNAME_FIELD = 'email'
     objects = MyUserManager()
